test: move inspector tests
